match_service/requirements.txt
django
channels[daphne]
channels_redis
djangorestframework
psycopg2>=2.9
pytest
pytest-django
pytest-asyncio
requests
pytest-mock
